refactor: Extract some functions to utils and add tests
