+++ src/composables/modeStates/animationGraph/utils.ts
@@ -32,11 +32,9 @@ import { mapFilter, mapReduce, toList } from '/@/utils/commons'
 import { getCtrlOrMetaStr } from '/@/utils/devices'
 import { gridRound } from '/@/utils/geometry'
 import {
-  cleanEdgeGenericsGroupAt,
   duplicateNodes,
   GetGraphNodeModule,
   isolateNodes,
-  updateInputConnection,
   validateConnection,
 } from '/@/utils/graphNodes'
 import {
@@ -99,62 +97,6 @@ export function getGridRoundedDiff(p: IVec2): IVec2 {
   return sub(getGridRoundedPoint(p), p)
 }
 
-export function updateNodeInput(
-  getGraphNodeModule: GetGraphNodeModule,
-  nodeMap: IdMap<GraphNode>,
-  inputId: string,
-  inputKey: string,
-  outputId: string,
-  outputKey: string
-): IdMap<GraphNode> {
-  if (!nodeMap[outputId] || !nodeMap[inputId]) return {}
-
-  const updated = updateInputConnection(
-    { node: nodeMap[outputId], key: outputKey },
-    { node: nodeMap[inputId], key: inputKey }
-  )
-  if (!updated) return {}
-
-  return {
-    [updated.id]: updated,
-    ...cleanEdgeGenericsGroupAt(
-      getGraphNodeModule,
-      { ...nodeMap, [updated.id]: updated },
-      { id: updated.id, key: inputKey }
-    ),
-  }
-}
-
-export function updateMultipleNodeInput(
-  getGraphNodeModule: GetGraphNodeModule,
-  nodeMap: IdMap<GraphNode>,
-  outputId: string,
-  outputKey: string,
-  inputs: GraphEdgeConnection[]
-): IdMap<GraphNode> {
-  if (inputs.length === 0) return {}
-
-  const updatedIds = new Set<string>()
-  const latestMap = { ...nodeMap }
-
-  inputs.forEach((input) => {
-    const updated = updateNodeInput(
-      getGraphNodeModule,
-      nodeMap,
-      input.nodeId,
-      input.key,
-      outputId,
-      outputKey
-    )
-    for (const id in updated) {
-      latestMap[id] = updated[id]
-      updatedIds.add(id)
-    }
-  })
-
-  return mapFilter(latestMap, (_, id) => updatedIds.has(id))
-}
-
 export function parseNodeEdgeInfo(target: ModeEventTarget): {
   key: string
   id: string
@@ -186,6 +128,7 @@ export const COMMAND_EXAM_SRC = {
   delete: { command: 'x', title: 'Delete' },
   grab: { command: 'g', title: 'Grab' },
   duplicate: { command: 'D', title: 'Duplicate' },
+  reconnect: { command: `${getCtrlOrMetaStr()} + Drag`, title: 'Reconnect' },
   cutEdge: { command: `${getCtrlOrMetaStr()} + R-Drag`, title: 'Disconnect' },
   clip: { command: `${getCtrlOrMetaStr()} + c`, title: 'Clip' },
   paste: { command: `${getCtrlOrMetaStr()} + v`, title: 'Paste' },
